+++ bin/jsdoc-to-ts.js
@@ -13,7 +13,7 @@ import jsdocToTs from '../index.js';
 process.on('uncaughtException', err => console.error(err.stack));
 
 function displayHelp () {
-	let e = 'node ' + path.relative(process.cwd(), __filename);
+	let e = 'node ' + path.relative(process.cwd(), __filename);	// eslint-disable-line no-undef
 	if (require.main !== module) e = 'jsdoc-to-ts';
 
 	console.log('  jsdoc-to-ts v' + require('../package.json').version);