+++ src/renderers.js
@@ -284,11 +284,11 @@ export function defaultHocRenderer ({section, imports, typeRenderer = renderType
 export function defaultComponentRenderer ({section, renderer, imports, typeRenderer = renderType}) {
 	const props = section.members.instance.filter(member => !member.kind);
 	const funcs = section.members.instance.filter(member => member.kind === 'function');
-
 	const omits = section.tags.reduce((res, tag) => tag.title === 'omit' ? res.concat(tag.description) : res, []);
 	const propsBase = calcPropsBaseName({imports, section});
 	const propsInterfaceName = `${section.name}Props`;
 	const propsInterface = renderInterface(propsInterfaceName, props, propsBase, typeRenderer, imports, omits);
+	let staticMembersDescriptionArray = [], staticMembersDescription = '';
 
 	imports.add({
 		module: 'react',
@@ -296,6 +296,14 @@ export function defaultComponentRenderer ({section, renderer, imports, typeRende
 		all: true
 	});
 
+	if (section.members.static.length > 0) {
+		for (let i = 0; i < section.members.static.length; i++) {
+			const tagName = section.members.static[i].tags.reduce((res, tag) => tag.title === 'name' ? res + tag.name : res, []);
+			const linkName = section.members.static[i].description.children[0].children.reduce((res, tag) => tag.type === 'link' ? res + tag.url.split('.')[1] : res, []);
+			staticMembersDescriptionArray.push(renderDescription(section.members.static[i]) + linkName + ' : ' + tagName + '\n\n');
+		}
+		staticMembersDescription = staticMembersDescriptionArray.reduce((acc, current) => acc + current + '\n');
+	}
 	return `${propsInterface}
 		${renderDescription(section)}
 		export class ${section.name} extends React.Component<Merge<React.HTMLProps<HTMLElement>, ${propsInterfaceName}>> {
@@ -304,6 +312,7 @@ export function defaultComponentRenderer ({section, renderer, imports, typeRende
 					.filter(Boolean)
 					.join('\n')
 			}
+			${staticMembersDescription}
 		}
 	`;
 }