+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Added the rendering of static members for a component.
 * Fixed security vulnerabilities.
+* Updated `jsonata` dependency
 
 ## [0.1.2] (May 31, 2022)