+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 ## [unreleased]
 
 * Migrate project to ESM.
+* Updated `jsonata` dependency to version 2.x and refactored code to use async functions.
 
 ## [0.1.3] (December 22, 2022)
 
+* Added the rendering of static members for a component.
 * Fixed security vulnerabilities.
 
 ## [0.1.2] (May 31, 2022)