+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## [0.1.3] (December 22, 2022)
 
+* Added the rendering of static members for a component.
 * Fixed security vulnerabilities.
 
 ## [0.1.2] (May 31, 2022)