+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Change Log
 
-## [0.1.3] (December 22, 2022)
+## [0.1.4] (March 2, 2023)
 
 * Added the rendering of static members for a component.
+* Fixed security vulnerabilities.
+
+## [0.1.3] (December 22, 2022)
+
 * Fixed security vulnerabilities.
 
 ## [0.1.2] (May 31, 2022)